JobID -> RID

+++ pipeline/build.sh
@@ -13,11 +13,13 @@ if [ "$distro" -ne 0 ]; then
     export ONGCP="true"
     export JAVA_HOME=/usr/lib/jvm/java-8-openjdk-amd64
     export PATH="$JAVA_HOME/bin:$PATH"
+    export LD_LIBRARY_PATH="/tmp/blast:."
 #    JAVA_INC="-I$JAVA_HOME/include"
 else
     echo "Building at NCBI on CentOS 7"
     export ONGCP="false"
     export JAVA_HOME=$(dirname $(dirname $(readlink -f $(which javac))))
+    export LD_LIBRARY_PATH="/tmp/blast:.:/opt/ncbi/gcc/4.9.3/lib64/"
 fi
 JAVA_INC=" -I$JAVA_HOME/include -I$JAVA_HOME/include/linux"
 export CLASSPATH="."
@@ -27,7 +29,6 @@ HDR="gov_nih_nlm_ncbi_blastjni_BlastJNI.h"
 rm -f $HDR
 rm -f BlastJNI.class
 rm -f blastjni.o
-#rm -rf blast-libs
 rm -rf gov
 rm -f *test.result
 rm -rf target
@@ -38,11 +39,61 @@ rm -f db_partitions.json db_partitions.jsonl
 
 #curl -o blast4spark.cpp https://svn.ncbi.nlm.nih.gov/viewvc/toolkit/branches/blast_gcp/src/algo/blast/api/blast4spark.cpp?view=co&content-type=text%2Fplain
 #curl -o blast4spark.hpp https://svn.ncbi.nlm.nih.gov/viewvc/toolkit/branches/blast_gcp/include/algo/blast/api/blast4spark.hpp?view=co&content-type=text%2Fplain
+
+echo "Compiling BlastJNI.java"
+mkdir -p gov/nih/nlm/ncbi
+javac -d gov/nih/nlm/ncbi -h . BlastJNI.java
+#rm -rf gov
+echo "/*" >> $HDR
+echo "$USER" >> $HDR
+javac -version >> $HDR 2>&1
+g++ --version | head -1 >> $HDR
+#date >> $HDR
+echo "*/" >> $HDR
+# javac
+#javac -d gov/nih/nlm/ncbi BlastJNI.java
+javac -d . BlastJNI.java
+#cp BlastJNI.class gov/nih/nlm/ncbi
+
 if [ "$ONGCP" = "false" ]; then
-    rm -f test_blast
+    echo "Compiling blastjni.cpp"
+    rm -f blastjni.so
+    g++ blastjni.cpp \
+        -L./int/blast/libs \
+        -std=gnu++11 \
+        -Wall -g  -I . \
+        -shared \
+        -fPIC \
+        $JAVA_INC \
+        -I /netopt/ncbi_tools64/c++.stable/include \
+        -I /netopt/ncbi_tools64/c++.stable/GCC493-ReleaseMT/inc \
+        -L /netopt/ncbi_tools64/c++.stable/GCC493-ReleaseMT/lib \
+        -L . \
+	-L ext \
+        -fopenmp -lxblastformat -lalign_format -ltaxon1 -lblastdb_format -lgene_info -lxformat -lxcleanup -lgbseq -lmlacli -lmla -lmedlars -lpubmed -lvalid -ltaxon3 -lxalnmgr -lblastxml -lblastxml2 -lxcgi -lxhtml -lproteinkmer -lxblast -lxalgoblastdbindex -lcomposition_adjustment -lxalgodustmask -lxalgowinmask -lseqmasks_io -lseqdb -lblast_services -lxalnmgr -lxobjutil -lxobjread -lvariation -lcreaders -lsubmit -lxnetblastcli -lxnetblast -lblastdb -lscoremat -ltables -lxregexp -lncbi_xloader_genbank -lncbi_xreader_id1 -lncbi_xreader_id2 -lncbi_xreader_cache -lncbi_xreader_pubseqos -ldbapi_driver -lncbi_xreader -lxconnext -lxconnect -lid1 -lid2 -lxobjmgr -lgenome_collection -lseqedit -lseqsplit -lsubmit -lseqset -lseq -lseqcode -lsequtil -lpub -lmedline -lbiblio -lgeneral -lxser -lxutil -lxncbi -lxcompress -llmdb -lpthread -lz -lbz2 -L/netopt/ncbi_tools64/lzo-2.05/lib64 -llzo2 -ldl -lz -lnsl -ldw -lrt -ldl -lm -lpthread\
+        -o blastjni.so
+    mkdir -p /tmp/blast
+    cp blastjni.so /tmp/blast
     cp -n /netopt/ncbi_tools64/lmdb-0.9.21/lib/*so ext
+    cp ext/liblmdb.so /tmp/blast/
+fi
+
+echo "Testing JNI"
+java -Djava.library.path=$PWD -cp . gov.nih.nlm.ncbi.blastjni.BlastJNI > test.result
+set +errexit
+CMP=$(cmp test.result test.expected)
+if [[ $? -ne 0 ]]; then
+    echo "Test failed"
+    sdiff -w 70 test.result test.expected
+    exit 1
+fi
+set -o errexit
+echo "Test OK"
+
+if [ "$ONGCP" = "false" ]; then
     echo "Compiling test_blast.cpp"
-    g++ test_blast.cpp -L./blast-libs \
+    rm -f test_blast
+    g++ test_blast.cpp -L./int/blast/libs \
         -std=gnu++11 \
         -Wall -g -fPIC -I . \
         -I /netopt/ncbi_tools64/c++.stable/include \
@@ -63,7 +114,6 @@ if [ "$ONGCP" = "false" ]; then
 else
     export BLASTDB=/tmp/blast/db
 fi
-export LD_LIBRARY_PATH="/tmp/blast:.:/opt/ncbi/gcc/4.9.3/lib64/"
 ./test_blast 1 \
 CCGCAAGCCAGAGCAACAGCTCTAACAAGCAGAAATTCTGACCAAACTGATCCGGTAAAACCGATCAACG \
 nt.04 blastn > blast_test.result
@@ -72,64 +122,6 @@ CMP=$(cmp blast_test.result blast_test.expected)
 if [[ $? -ne 0 ]]; then
     echo "Test failed"
     sdiff -w 70 blast_test.result blast_test.expected | head
-    exit 1
-fi
-set -o errexit
-echo "Test OK"
-
-echo "Compiling BlastJNI.java"
-mkdir -p gov/nih/nlm/ncbi
-javac -d gov/nih/nlm/ncbi -h . BlastJNI.java
-rm -rf gov
-echo "/*" >> $HDR
-echo "$USER" >> $HDR
-javac -version >> $HDR 2>&1
-g++ --version | head -1 >> $HDR
-#date >> $HDR
-echo "*/" >> $HDR
-
-
-# javac
-#javac -d gov/nih/nlm/ncbi BlastJNI.java
-javac -d . BlastJNI.java
-#cp BlastJNI.class gov/nih/nlm/ncbi
-
-if [ "$ONGCP" = "false" ]; then
-    echo "Compiling blastjni.cpp"
-    rm -f blastjni.so
-    g++ blastjni.cpp \
-        -L./blast-libs \
-        -std=gnu++11 \
-        -Wall -g  -I . \
-        -shared \
-        -fPIC \
-        $JAVA_INC \
-        -I /netopt/ncbi_tools64/c++.stable/include \
-        -I /netopt/ncbi_tools64/c++.stable/GCC493-ReleaseMT/inc \
-        -L /netopt/ncbi_tools64/c++.stable/GCC493-ReleaseMT/lib \
-        -L . \
-	-L ext \
-        -fopenmp -lxblastformat -lalign_format -ltaxon1 -lblastdb_format -lgene_info -lxformat -lxcleanup -lgbseq -lmlacli -lmla -lmedlars -lpubmed -lvalid -ltaxon3 -lxalnmgr -lblastxml -lblastxml2 -lxcgi -lxhtml -lproteinkmer -lxblast -lxalgoblastdbindex -lcomposition_adjustment -lxalgodustmask -lxalgowinmask -lseqmasks_io -lseqdb -lblast_services -lxalnmgr -lxobjutil -lxobjread -lvariation -lcreaders -lsubmit -lxnetblastcli -lxnetblast -lblastdb -lscoremat -ltables -lxregexp -lncbi_xloader_genbank -lncbi_xreader_id1 -lncbi_xreader_id2 -lncbi_xreader_cache -lncbi_xreader_pubseqos -ldbapi_driver -lncbi_xreader -lxconnext -lxconnect -lid1 -lid2 -lxobjmgr -lgenome_collection -lseqedit -lseqsplit -lsubmit -lseqset -lseq -lseqcode -lsequtil -lpub -lmedline -lbiblio -lgeneral -lxser -lxutil -lxncbi -lxcompress -llmdb -lpthread -lz -lbz2 -L/netopt/ncbi_tools64/lzo-2.05/lib64 -llzo2 -ldl -lz -lnsl -ldw -lrt -ldl -lm -lpthread\
-        -o blastjni.so
-    # TODO: llmdb-static?
-    mkdir -p /tmp/blast
-    cp blastjni.so /tmp/blast
-
-    if [ ! -d blast-libs ]; then
-        mkdir -p blast-libs
-        cd blast-libs
-        tar -xf ~/blast-libs.tgz
-        cd ..
-    fi
-fi
-
-echo "Testing JNI"
-java -Djava.library.path=$PWD -cp . gov.nih.nlm.ncbi.blastjni.BlastJNI > test.result
-set +errexit
-CMP=$(cmp test.result test.expected)
-if [[ $? -ne 0 ]]; then
-    echo "Test failed"
-    sdiff -w 70 test.result test.expected
     exit 1
 fi
 set -o errexit
@@ -209,6 +201,10 @@ gcloud dataproc --region us-east4 \
     'gs://blastgcp-pipeline-test/scipts/cluster_initialize.sh' \
     --initialization-actions-timeout 60 # Default 10m \
     --max-age=8h
+
+    git clone https://github.com/ncbi/blast-gcp.git
+    cd blast-gcp
+    git checkout engineering
 
 # Can be useful for debugging
  export SPARK_PRINT_LAUNCH_COMMAND=1
